Boxes now throwable/stackable
Still working out some kinks. About to refactor the character controller to make diagonal movement not 2x the speed of cardinal movement.

Wish me luck.

+++ Assets/Script/Weapons/Pistol.cs
@@ -27,7 +27,7 @@ public class Pistol : Weapon {
         }
     }
 
-    public override float MaxAmmo {
+    public override int MaxAmmo {
         get {
             return maxAmmo;
         }
